% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVars.R
\name{addVars}
\alias{addVars}
\title{addVars}
\usage{
addVars(path_user, results_directoryName, csv_decimalSeparator,
  csv_columnSeparator, ErrorOccured)
}
\arguments{
\item{path_user}{this is path to users model directory containing results, data, and gis folders}

\item{results_directoryName}{this is the users directory name for the results}

\item{csv_decimalSeparator}{decimal separator for read/write csv files}

\item{csv_columnSeparator}{column separator for read/write csv files}

\item{ErrorOccured}{yes/no indicating if a previous error has occured \code{ErrorOccured<-"no"}}
}
\description{
Tests for sparrowNames found in parameters.csv, but not in dataDictionary.csv or design_matrix.csv.
edits dataDictionary.csv and/or design_matrix.csv adding missing sparrowName
and opens design_matrix.csv and userModifyData.R for edit.
}
