% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(path, input_data_fileName, csv_decimalSeparator, csv_columnSeparator,
  batch_mode, ErrorOccured)
}
\arguments{
\item{path}{this is users path_data automatically found}

\item{input_data_fileName}{this is the users fileName for the input data, if csv must include the file extension}

\item{batch_mode}{yes/no indicating whether or not batch processing is used}

\item{ErrorOccured}{yes/no indicating if a previous error has occured \code{ErrorOccured<-"no"}}
}
\description{
This reads the users input data (either .csv or binary format)
}
