% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeObjects.R
\name{removeObjects}
\alias{removeObjects}
\title{removeObjects}
\usage{
removeObjects(remove.list)
}
\arguments{
\item{remove.list}{a character vector of objects to be removed}
}
\description{
Remove objects from the global environment if the object exists.
Warnings will be suppressed.
}
\examples{
x<-c(1,2,3)
y<-data.frame(column1 = c(NA,4,5))
z<-"string"
removeObjects(c("x","y","z"))
}
\keyword{internal}
