% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceNAs.R
\name{replaceNAs}
\alias{replaceNAs}
\title{replaceNAs}
\usage{
replaceNAs(listColumns)
}
\arguments{
\item{listColumns}{A named list of variables in which to replace NAs.
Recommend using \code{named.list()} to generate a named list of variables.}
}
\description{
Replace all NAs with 0's. Output is saved to parent.frame().
}
\examples{
x<-c(1,2,3)
y<-c(NA,4,5)
replaceNAs(named.list(x,y))
}
\keyword{external}
